feat(vfs): add job for vfs (#2820)
* feat(vfs): add job for vfs

* feat(ios): add ios progress job for vfs

* feat(ios): add ios progress job for vfs

* feat(ios): add ios progress job for vfs

* feat(ios): add ios progress job for vfs

* feat(vfs): adapt to voltron

* feat(ios): add ios progress job for vfs

* feat(voltron): vfs logic compact

* fix(vfs): merge 3.0dev to vfs

+++ framework/android/src/main/java/com/tencent/mtt/hippy/bridge/HippyBridgeImpl.java
@@ -316,7 +316,7 @@ public class HippyBridgeImpl implements HippyBridge, JSBridgeProxy, DevRemoteDeb
                             }
 
                             @Override
-                            public void onFetchProgress(float total, float loaded) {
+                            public void onFetchProgress(long total, long loaded) {
                                 // Nothing need to do here.
                             }
                         });
