+++ framework/android/src/main/java/com/tencent/mtt/hippy/modules/nativemodules/image/ImageLoaderModule.java
@@ -84,7 +84,7 @@ public class ImageLoaderModule extends HippyNativeModuleBase {
                     }
 
                     @Override
-                    public void onFetchProgress(float total, float loaded) {
+                    public void onFetchProgress(long total, long loaded) {
                         // Nothing need to do here.
                     }
                 });
@@ -100,7 +100,7 @@ public class ImageLoaderModule extends HippyNativeModuleBase {
                     }
 
                     @Override
-                    public void onFetchProgress(float total, float loaded) {
+                    public void onFetchProgress(long total, long loaded) {
                         // Nothing need to do here.
                     }
                 });
