+++ renderer/native/android/src/main/java/com/tencent/renderer/component/image/ImageComponent.java
@@ -254,7 +254,7 @@ public class ImageComponent extends Component {
             }
 
             @Override
-            public void onRequestProgress(float total, float loaded) {
+            public void onRequestProgress(long total, long loaded) {
                 if (sourceType == ImageSourceType.SRC) {
                     onFetchImageProgress(total, loaded);
                 }
